feat(billing): refactor into multiple router
