+++ apps/api/src/auth/services/ability/ability.service.ts
@@ -24,11 +24,7 @@ export class AbilityService {
     REGULAR_USER: [
       { action: ["create", "read", "sign"], subject: "UserWallet", conditions: { userId: "${user.id}" } },
       { action: "read", subject: "User", conditions: { id: "${user.id}" } },
-      { action: "read", subject: "StripePrice" },
-      { action: ["create", "read", "delete"], subject: "PaymentMethod" },
-      { action: ["create", "read"], subject: "PaymentIntent" },
-      { action: ["create", "read"], subject: "Coupon" },
-      { action: "read", subject: "Transaction" },
+      { action: ["create", "read", "delete"], subject: "Payment" },
       { action: "create", subject: "VerificationEmail", conditions: { id: "${user.id}" } },
       { action: "manage", subject: "DeploymentSetting", conditions: { userId: "${user.id}" } },
       { action: "manage", subject: "Alert", conditions: { userId: "${user.id}" } },
@@ -37,11 +33,7 @@ export class AbilityService {
     REGULAR_PAYING_USER: [
       { action: ["create", "read", "sign"], subject: "UserWallet", conditions: { userId: "${user.id}" } },
       { action: "read", subject: "User", conditions: { id: "${user.id}" } },
-      { action: "read", subject: "StripePrice" },
-      { action: ["create", "read", "delete"], subject: "PaymentMethod" },
-      { action: ["create", "read"], subject: "PaymentIntent" },
-      { action: ["create", "read"], subject: "Coupon" },
-      { action: "read", subject: "Transaction" },
+      { action: ["create", "read", "delete"], subject: "Payment" },
       { action: "create", subject: "VerificationEmail", conditions: { id: "${user.id}" } },
       { action: "manage", subject: "DeploymentSetting", conditions: { userId: "${user.id}" } },
       { action: "manage", subject: "ApiKey", conditions: { userId: "${user.id}" } },
