+++ apps/api/src/billing/routes/index.ts
@@ -4,5 +4,7 @@ export * from "@src/billing/routes/checkout/checkout.router";
 export * from "@src/billing/routes/sign-and-broadcast-tx/sign-and-broadcast-tx.router";
 export * from "@src/billing/routes/stripe-webhook/stripe-webhook.router";
 export * from "@src/billing/routes/stripe-prices/stripe-prices.router";
-export * from "@src/billing/routes/stripe-setup/stripe-setup.router";
+export * from "@src/billing/routes/coupons/coupons.router";
+export * from "@src/billing/routes/transactions/transactions.router";
+export * from "@src/billing/routes/payment-methods/payment-methods.router";
 export * from "@src/billing/routes/get-balances/get-balances.router";
