+++ packages/http-sdk/src/stripe/stripe.types.ts
@@ -1,0 +1,53 @@
+export interface StripePrice {
+  unitAmount?: number;
+  isCustom: boolean;
+  currency: string;
+}
+
+export interface SetupIntentResponse {
+  clientSecret: string;
+}
+
+export interface PaymentResponse {
+  error?: {
+    message: string;
+  };
+}
+
+export interface Coupon {
+  id: string;
+  percent_off?: number | null;
+  amount_off?: number | null;
+  valid: boolean;
+  name?: string;
+  description?: string;
+}
+
+export interface CouponResponse {
+  coupon: Coupon;
+}
+
+export interface CustomerDiscountsResponse {
+  discounts: Array<{
+    type: "coupon" | "promotion_code";
+    id: string;
+    name?: string;
+    code?: string;
+    percent_off?: number;
+    amount_off?: number;
+    currency?: string;
+    valid: boolean;
+  }>;
+}
+
+export interface Charge {
+  id: string;
+  amount: number;
+  currency: string;
+  status: string;
+  created: number;
+  paymentMethod: any;
+  receiptUrl?: string;
+  description?: string;
+  metadata?: Record<string, string>;
+}
